refactor: move DocumentSymbolAdapter to new file
